requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn>=0.32.0
lark-oapi>=1.4.0
anthropic>=0.52.0
claude-agent-sdk>=0.1.60
python-dotenv>=1.0.0
aiosqlite>=0.20.0
clickhouse-driver>=0.2.9
apscheduler>=3.10.0
jinja2>=3.1.0
itsdangerous>=2.1.0
python-multipart>=0.0.9
